updated getAuthenticatedUserId method and fixed code style

+++ Utils.php
@@ -2003,20 +2003,21 @@ class Utils
     }
 
     /**
-     * Summary of parse_size
+     * Get size parameters from php.ini
      * @param string $sizeName
      * @return float size in bytes
      */
-    public static function getSizeFromIni($sizeName) {
+    public static function getSizeFromIni($sizeName)
+    {
         $size = ini_get($sizeName);
         if ($size) {
             $unit = preg_replace('/[^bkmgtpezy]/i', '', $size); // Remove the non-unit characters from the size.
-            $size = preg_replace('/[^0-9\.]/', '', $size); // Remove the non-numeric characters from the size.
+            $size = (int) preg_replace('/[^0-9\.]/', '', $size); // Remove the non-numeric characters from the size.
             if ($unit) {
-            // Find the position of the unit in the ordered string which is the power of magnitude to multiply a kilobyte by.
-            return round($size * pow(1024, stripos('bkmgtpezy', $unit[0])));
+                // Find the position of the unit in the ordered string which is the power of magnitude to multiply a kilobyte by.
+                return round($size * pow(1024, stripos('bkmgtpezy', $unit[0])));
             } else {
-            return round($size);
+                return round($size);
             }
         } else {
             return 0;
